requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
numpy==1.24.2
pandas==1.5.3
scikit_learn==1.2.1
scipy==1.11.2
xgboost==1.7.4
Werkzeug==2.3.7
pyyaml==6.0.1
